Controlled on cascade delete according with functionality. Changed client port to 4200. Skinny Category controller.

+++ src/main/java/com/dam/kanpeki/service/impl/CategoryServiceImpl.java
@@ -8,7 +8,12 @@ import org.springframework.data.domain.Example;
 import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.stereotype.Service;
 
+import com.dam.kanpeki.exception.DataNotFoundException;
+import com.dam.kanpeki.exception.InvalidOperationOnCategoryException;
 import com.dam.kanpeki.model.Category;
+import com.dam.kanpeki.model.dto.RequestCategoryDTO;
+import com.dam.kanpeki.model.dto.ResponseCategoryDTO;
+import com.dam.kanpeki.model.dto.mapper.CategoryDTOMapperStruct;
 import com.dam.kanpeki.repository.CategoryRepository;
 import com.dam.kanpeki.service.CategoryServiceI;
 
@@ -18,9 +23,13 @@ public class CategoryServiceImpl implements CategoryServiceI {
 	@Autowired
 	private CategoryRepository catRepo;
 
+	@Autowired
+	private CategoryDTOMapperStruct mapper;
+
 	@Override
-	public List<Category> findAllCategories() {
-		return catRepo.findAll();
+	public List<ResponseCategoryDTO> findAllCategories() {
+		List<Category> catList = catRepo.findAll();
+		return mapper.toCategoryDTOList(catList.stream());
 	}
 
 	@Override
@@ -34,7 +43,7 @@ public class CategoryServiceImpl implements CategoryServiceI {
 	}
 
 	@Override
-	public List<Category> findCategoriesByMatcher(String wField) {
+	public List<ResponseCategoryDTO> findCategoriesByMatcher(String wField) {
 		Category cat = new Category();
 		cat.setUnitName(wField);
 		cat.setCategoryName(wField);
@@ -45,27 +54,56 @@ public class CategoryServiceImpl implements CategoryServiceI {
 
 		Example<Category> catExample = Example.of(cat, customExMatcher);
 
-		return catRepo.findAll(catExample);
+		return mapper.toCategoryDTOList(catRepo.findAll(catExample).stream());
 	}
 
 	@Override
-	public Optional<Category> findById(Long id) {
-		return catRepo.findById(id);
+	public Optional<ResponseCategoryDTO> findById(Long id) {
+		Optional<Category> opCat = catRepo.findById(id);
+		return Optional.of(mapper.toCategoryDTO(opCat.get()));
 	}
 
 	@Override
-	public Category addWord(Category cat) {
-		return catRepo.save(cat);
+	public ResponseCategoryDTO addWord(RequestCategoryDTO cat) {
+
+		return mapper.toCategoryDTO(catRepo.save(mapper.requestCategoryDTOtoCategory(cat)));
 	}
 
 	@Override
 	public void removeCategoryById(Long id) {
+		Optional<Category> opCat = catRepo.findById(id);
+
+		if ((!opCat.get().getWords().isEmpty()) || (!opCat.get().getQuestions().isEmpty())
+				|| (!opCat.get().getResults().isEmpty())) {
+
+			throw new InvalidOperationOnCategoryException(
+					(!opCat.get().getWords().isEmpty() ? " Words ".concat(String.valueOf(opCat.get().getWords().size()))
+							: "")
+							+ (!opCat.get().getWords().isEmpty()
+									? " Questions ".concat(String.valueOf(opCat.get().getWords().size()))
+									: "")
+							+ (!opCat.get().getWords().isEmpty()
+									? " Results ".concat(String.valueOf(opCat.get().getWords().size()))
+									: ""));
+		}
+
 		catRepo.deleteById(id);
 	}
 
 	@Override
-	public void updateCategory(Category cat) {
-		catRepo.save(cat);
+	public ResponseCategoryDTO updateCategory(RequestCategoryDTO cat, Long id) {
+		Category mappedCat = mapper.requestCategoryDTOtoCategory(cat);
+
+		Category mappedCatUpdated = catRepo.findById(id).map(newCat -> {
+			newCat.setUnitName(mappedCat.getUnitName());
+			newCat.setCategoryName(mappedCat.getCategoryName());
+			newCat.setIsQuestion(mappedCat.getIsQuestion());
+			catRepo.save(newCat);
+			return newCat;
+		}).orElseThrow(() -> new DataNotFoundException(""));
+
+		return mapper.toCategoryDTO(mappedCatUpdated);
+
 	}
 
 }
